Transferred data_manager.py to database.py
